feat: implemented the validate methods into githubinfo.py
